feat!: inlined schema definitions inside the KafkaSchema custom resource
v1beta1 approach was based on the convention that the actual schema
(payload/structure) was stored in configmap and only referenced by
KafkaSchema resource. There are a few problems related to that approach:

1. inconsistent behaviour - when you create a schema resource, configmap
needs to be already applied in your namespace. but finalizer deletes
that configmap during cleanup
2. race condition between K8S resources - during reconciliation, it's
assumed that configmap was updated prior to updating KafkaSchema. If it
wasn't, changes will not be applied. On the other hand, configmap has to
be deleted after the custom resource, otherwise finalizer will fail.
3. difficulties in managing "independent" K8S resources in a single
chart - it's not easy/recommended to use owner references from within a
single chart, and playing with finalizers to "block deletion" of
configmap is tricky

Instead, I would like to propose a simpler approach, where the entire
KafkaSchema resource is self-contained. You can define your schema
payload directly in the resource or use helm templating to populate it
from an external file

BREAKING CHANGE: CR no longer supports reference to schema configmap
through spec.data.configRef